requirements.txt
streamlit
google-generativeai
python-dotenv
psycopg2-binary
reportlab
pypdf
fpdf2